token

+++ js/item.js
@@ -1,5 +1,20 @@
 $(document).ready(function () {
-    const url = 'http://172.34.96.144:8000'
+    const url = 'http://172.34.11.117:8000'
+    const getToken = () => {
+        const token = sessionStorage.getItem('token');
+
+        if (!token) {
+            Swal.fire({
+                icon: 'warning',
+                text: 'You must be logged in to access this page.',
+                showConfirmButton: true
+            }).then(() => {
+                window.location.href = 'login.html';
+            });
+            return;
+        }
+        return JSON.parse(token)
+    }
 
     $('#itable').DataTable({
         ajax: {
@@ -109,6 +124,9 @@ $(document).ready(function () {
             method: "GET",
             url: `${url}/api/v1/items/${id}`,
             dataType: "json",
+            headers: {
+                "Authorization": "Bearer " + getToken()
+            },
             success: function (data) {
                 const { description,
                     item_id,
@@ -149,6 +167,9 @@ $(document).ready(function () {
             processData: false,
 
             dataType: "json",
+            headers: {
+                "Authorization": "Bearer " + getToken()
+            },
             success: function (data) {
                 console.log(data);
                 $('#itemModal').modal("hide");
@@ -167,48 +188,48 @@ $(document).ready(function () {
         var id = $(this).data('id');
         var $row = $(this).closest('tr');
         console.log(id);
-        // if (getToken()) {
-        bootbox.confirm({
-            message: "do you want to delete this item",
-            buttons: {
-                confirm: {
-                    label: 'yes',
-                    className: 'btn-success'
+        if (getToken()) {
+            bootbox.confirm({
+                message: "do you want to delete this item",
+                buttons: {
+                    confirm: {
+                        label: 'yes',
+                        className: 'btn-success'
+                    },
+                    cancel: {
+                        label: 'no',
+                        className: 'btn-danger'
+                    }
                 },
-                cancel: {
-                    label: 'no',
-                    className: 'btn-danger'
-                }
-            },
-            callback: function (result) {
-                console.log(result);
-                if (result) {
-                    $.ajax({
-                        method: "DELETE",
-                        url: `${url}/api/v1/items/${id}`,
-                        dataType: "json",
-                        // headers: {
-                        //     "Authorization": "Bearer " + getToken()
-                        // },
-                        success: function (data) {
-                            console.log(data);
-                            $row.fadeOut(4000, function () {
-                                table.row($row).remove().draw();
-                            });
+                callback: function (result) {
+                    console.log(result);
+                    if (result) {
+                        $.ajax({
+                            method: "DELETE",
+                            url: `${url}/api/v1/items/${id}`,
+                            dataType: "json",
+                            headers: {
+                                "Authorization": "Bearer " + getToken()
+                            },
+                            success: function (data) {
+                                console.log(data);
+                                $row.fadeOut(4000, function () {
+                                    table.row($row).remove().draw();
+                                });
 
-                            bootbox.alert(data.success);
-                        },
-                        error: function (error) {
-                            bootbox.alert(data.error);
-                        }
-                    });
+                                bootbox.alert(data.success);
+                            },
+                            error: function (error) {
+                                bootbox.alert(data.error);
+                            }
+                        });
+
+                    }
 
                 }
+            });
 
-            }
-        });
-
-        // }
+        }
 
     })
 })
